Extract indexation to a separate module
